Añadir libro, lista de libros

+++ src/main/java/cl/andres/java/biblioteca/repository/AutorRepository.java
@@ -2,20 +2,18 @@ package cl.andres.java.biblioteca.repository;
 
 import java.util.List;
 
-import org.springframework.stereotype.Repository;
-
 import cl.andres.java.biblioteca.model.Autor;
 
 public interface AutorRepository {
 
 	public List<Autor> findAll();
 	
-	public Autor findById(int id);
+	public Autor findById(Long id);
 	
 	public void create(Autor autor);
 	
 	public void edit(Autor autor);
 	
-	public void delete(int id);
+	public void delete(Long id);
 	
 }
